Add compiler warning options and update various source files for improved safety and clarity
- Introduced TD_ENABLE_WERROR option in CMakeLists.txt to treat compiler warnings as errors.
- Updated multiple source files to use safer functions and improved type handling.
- Adjusted log timestamps for accurate testing duration.
- Refactored code to enhance readability and maintainability.

+++ src/layer.cpp
@@ -23,7 +23,7 @@ void LayerClass::Sort() {
 }
 
 int LayerClass::Sorted_Add(ObjectClass const* object) {
-  if (ActiveCount >= Length()) {
+  if (ActiveCount >= static_cast<int>(Length())) {
     if ((IsAllocated || !VectorMax) && GrowthStep > 0) {
       if (!Resize(Length() + GrowthStep)) {
         return false;
